end-to-end flow from unity to lamda back to unity working
